+++ bpdm-pool/src/test/kotlin/org/eclipse/tractusx/bpdm/pool/v6/operator/metadata/LegalFormIT.kt
@@ -26,7 +26,6 @@ import org.eclipse.tractusx.bpdm.common.dto.PaginationRequest
 import org.eclipse.tractusx.bpdm.pool.api.v6.model.LegalFormDto
 import org.eclipse.tractusx.bpdm.pool.api.v6.model.request.LegalFormRequest
 import org.eclipse.tractusx.bpdm.pool.v6.operator.OperatorTest
-import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 import org.springframework.web.reactive.function.client.WebClientResponseException
 
@@ -107,13 +106,8 @@ class LegalFormIT: OperatorTest(){
     /**
      * WHEN operator tries to create legal form for unknown administrative area
      * THEN 400 BAD REQUEST returned
-     *
-     * ToDo:
-     * Currently not fulfilled!
-     * Bug Ticket for tracking: https://github.com/eclipse-tractusx/bpdm/issues/1446
      */
     @Test
-    @Disabled
     fun `try create new legal form with unknown admin area`(){
         val request = createLegalFormRequest(testName)
             .copy(administrativeAreaLevel1 = "UNKNOWN")